Add failing test scenario for a shopping cart with offers

+++ src/main/scala/Products.scala
@@ -1,6 +1,6 @@
 object Products {
-  val prices: Map[String, Double] = Map(
-    "Apple" -> 0.60,
-    "Orange" -> 0.25
+  val prices: Map[String, BigDecimal] = Map(
+    "Apple" -> BigDecimal(0.60),
+    "Orange" -> BigDecimal(0.25)
   )
 }
